Fix broken submodule path
When I changed the OpenAL Soft submodule to point to a fork, I forgot to
change the corresponding repository path in .gitmodules. Because of
this, GitHub (and probably command-line Git as well) can't find the
commit the submodule is pointing to.

This commit updates the submodule path to what it should have been. The
only reason this commit doesn't amend the previous one is because I
had pushed it already.